change filter location
