+++ Obsidian/Shared/WadFilter.razor
@@ -28,7 +28,8 @@
                 <MudChip Class="rounded ma-0" Size="Size.Small" Text="F" />
             </MudStack>
         </MudStack>
-        <MudTextField T="string" @ref="InputField"
+        <MudTextField Margin="Margin.Dense"
+                      T="string" @ref="InputField"
                       Class="mt-0 mud-elevation-2"
                       Variant="Variant.Filled"
                       DebounceInterval="500"
